- Image add feature - Multiple delete feature

+++ database/migrations/2019_02_07_093051_create_products_table.php
@@ -18,8 +18,8 @@ class CreateProductsTable extends Migration
             $table->string('name', 40);
             $table->string('sku');
             $table->boolean('status');
-            $table->float('basePrice');
-            $table->float('specialPrice')->nullable();
+            $table->double('basePrice');
+            $table->double('specialPrice')->nullable();
             $table->string('image')->nullable();
             $table->longText('description');
             $table->timestamps();
